requirements.txt
Flask
torch
Pillow
